+++ src/Model/Gunner.java
@@ -1,7 +1,7 @@
 package Model;
 
 public class Gunner extends Dwarf{
-    public Gunner(String theName, double theX, double theY, int theHealth, double theWidth, double theHeight, double theMoveSpeed, Weapon theWeapon) {
+    public Gunner(String theName, double theX, double theY, int theHealth, int theWidth, int theHeight, double theMoveSpeed, Weapon theWeapon) {
         super(theName, theX, theY, theHealth, theWidth, theHeight, theMoveSpeed, theWeapon);
     }
 }
